fix(tearchan): Change size type to usize from u32

+++ src/math/change_range.rs
@@ -2,13 +2,13 @@ use std::ops::Range;
 
 #[derive(Debug)]
 pub struct ChangeRange {
-    pub size: u32,
-    range_start: u32,
-    range_end: u32,
+    pub size: usize,
+    range_start: usize,
+    range_end: usize,
 }
 
 impl ChangeRange {
-    pub fn new(size: u32) -> ChangeRange {
+    pub fn new(size: usize) -> ChangeRange {
         ChangeRange {
             size,
             range_start: 0,
@@ -16,7 +16,7 @@ impl ChangeRange {
         }
     }
 
-    pub fn resize(&mut self, size: u32) {
+    pub fn resize(&mut self, size: usize) {
         self.size = size;
         self.range_end = size;
         if self.range_start > size {
@@ -24,15 +24,15 @@ impl ChangeRange {
         }
     }
 
-    pub fn resize_and_update(&mut self, start: u32, size: u32) {
+    pub fn resize_and_update(&mut self, start: usize, size: usize) {
         self.size = size;
         self.update(start, size);
     }
 
-    pub fn update(&mut self, start: u32, end: u32) {
+    pub fn update(&mut self, start: usize, end: usize) {
         debug_assert!(start < end, "start wasn't less than end");
 
-        if self.range_start == std::u32::MAX {
+        if self.range_start == std::usize::MAX {
             self.range_start = start;
             self.range_end = end;
             return;
@@ -51,12 +51,12 @@ impl ChangeRange {
     }
 
     pub fn reset(&mut self) {
-        self.range_start = std::u32::MAX;
-        self.range_end = std::u32::MAX;
+        self.range_start = std::usize::MAX;
+        self.range_end = std::usize::MAX;
     }
 
-    pub fn get_range(&self) -> Option<Range<u32>> {
-        if self.range_start == std::u32::MAX {
+    pub fn get_range(&self) -> Option<Range<usize>> {
+        if self.range_start == std::usize::MAX {
             return None;
         }
         Some(Range {
